farmstack-backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
APScheduler==3.10.4
asgiref==3.6.0
astroid==2.13.2
async-timeout==4.0.3
asyncio==3.4.3
attrs==21.4.0
backoff==2.2.1
beautifulsoup4==4.12.3
billiard==4.2.0
CacheControl==0.12.11
cachetools==5.3.3
cachy==0.3.0
celery==5.3.6
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cleo==0.8.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
clikit==0.6.2
configobj==5.0.8
coreapi==2.3.3
coreschema==0.0.4
coverage==7.0.3
coloredlogs==15.0.1
contourpy==1.2.1
crashtest==0.3.1
cryptography
cycler==0.12.1
cssutils==2.6.0
dataclasses-json==0.6.7
dateutils==0.6.12
decorator==4.4.2
deprecation==2.1.0
deepdiff==7.0.1
dill==0.3.6
distlib==0.3.4
Django==4.1.5
django-apscheduler==0.6.2
django-configurations==2.3.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-extensions==3.1.5
django-filter==22.1
django-nose==1.4.7
django-rest-framework-braces==0.3.4
django_debug_toolbar==3.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
drf-generators==0.5.0
drf-spectacular==0.22.1
drf-spectacular-sidecar==2022.7.1
drf-yasg==1.20.0
effdet==0.4.1
emoji==2.12.1
et-xmlfile==1.1.0
exceptiongroup==1.1.0
factory-boy==3.2.1
Faker==16.6.0
ffmpeg==1.4
filelock==3.15.4
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.6.1
flower==2.0.1
frozenlist==1.4.0
future==0.18.3
google-api-core==2.19.1
google-api-python-client==2.119.0
google-auth==2.28.1
google-cloud-vision==3.7.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
html5lib==1.1
httplib2==0.22.0
humanize==4.9.0
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
iopath==0.1.10
imageio==2.33.0
imageio-ffmpeg==0.4.9
importlib-metadata==6.0.0
importlib-resources==5.10.2
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.1.4
joblib==1.4.2
jsonpath-python==1.0.6
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.17.3
keyring==23.13.1
kiwisolver==1.4.5
kombu==5.3.5
langchain==0.1.1
layoutparser==0.3.4
langchain-community==0.0.20
langchain-core==0.1.23
langchainplus-sdk==0.0.16
langdetect==1.0.9
langsmith==0.0.87
lazy-object-proxy==1.7.1
lockfile==0.12.2
lxml==5.2.2
Markdown==3.6
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.2
mpmath==1.3.0
mccabe==0.7.0
model-bakery==1.9.0
more-itertools==9.0.0
moviepy==1.0.3
msgpack==1.0.4
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.0.28
nest-asyncio==1.6.0
nltk==3.8.1
networkx==3.3
nose==1.3.7
numpy==1.26.4
olefile==0.47
omegaconf==2.3.0
onnx==1.16.1
opencv-python==4.10.0.84
oauthlib==3.2.2
openai==1.30.3
openpyxl==3.1.5
ordered-set==4.1.0
orjson==3.9.15
packaging==23.2
pandas==2.2.2
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.1
pikepdf==9.0.0
pastel==0.2.1
pathspec==0.9.0
pexpect==4.8.0
pgvector==0.2.4
phonenumbers==8.13.15
pillow==10.4.0
pillow_heif==0.17.0
pkginfo==1.8.2
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
plazy==0.1.5
pluggy==1.0.0
proglog==0.1.10
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf
psycopg==3.1.14
psutil==6.0.0
psycopg-binary==3.1.18
psycopg2-binary
# psycopg
ptyprocess==0.7.0
py==1.11.0
pyaml==21.10.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.8
pycparser==2.22
pydantic==1.10.14
pypandoc==1.13
pydantic_core==2.16.3
PyJWT==2.4.0
pylev==1.4.0
pylint==2.15.10
PyMuPDF==1.23.12
PyMuPDFb==1.23.9
pyotp==2.6.0
pypdf==4.2.0
pypdfium2==4.30.0
pytesseract==0.3.10
pyparsing==3.1.2
pyrsistent==0.19.3
pytest==7.2.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.0
python-gnupg==0.5.1
python-http-client==3.3.7
python-iso639==2024.4.27
python-ldap==3.4.4
python-oxmsg==0.0.1
python-multipart==0.0.9
python-magic==0.4.27
python-on-whales==0.55.0
python-pptx==0.6.23
pytube==15.0.0
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.4
redis==5.0.1
regex==2024.5.15
reportlab==4.1.0
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.8
safetensors==0.4.3
# secrets
scipy==1.14.0
scikit-learn==1.5.1
SecretStorage==3.3.2
sendgrid==6.9.7
shellingham==1.4.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.27
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
systematic==4.8.7
tabulate==0.9.0
tenacity==8.2.3
testcontainers==3.7.1
tiktoken==0.5.2
timm==1.0.7
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.11.0
torch==2.3.1
torchvision==0.18.1
tornado==6.4
tqdm==4.66.4
transformers==4.42.3
typer==0.6.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
unstructured==0.14.9
unstructured-client==0.23.8
unstructured-inference==0.7.36
unstructured.pytesseract==0.3.12
urllib3==2.2.2
uritemplate==4.1.1
vine==5.1.0
virtualenv==20.14.1
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.7.0
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.2.0
xlwt==1.3.0
xmltodict==0.13.0
yarl==1.9.3
zipp==3.11.0
qdrant-client==1.9.1
boto3
django-storages
yt-dlp
azure-storage-blob
azure-core
dropbox
django_celery_beat
jq